basic functions and styling done
